Fix issues #58, #55: Integer overflow and config validation
**Issue #58 - FileSystemVolumeManager.getVolumeSizeLimit() overflow:**
- Fixed integer overflow when converting MB to bytes for large volumes
- Cast maxSizeMB to long before multiplication
- Prevents overflow for volumes > ~2GB (previously produced negative values)

**Issue #55 - ApiServerConfig.Builder validation:**
- Added validation that apiKey is set when enableAuth is true
- Prevents invalid configuration where authentication is enabled but will always fail
- Fail fast with clear error message

Both fixes prevent silent data corruption and invalid configurations.

+++ jplatform-storage/src/main/java/org/flossware/jplatform/storage/FileSystemVolumeManager.java
@@ -165,7 +165,7 @@ public class FileSystemVolumeManager implements VolumeManager {
         if (volume == null) {
             throw new IllegalArgumentException("Volume not defined: " + volumeName);
         }
-        return volume.getMaxSizeMB() * 1024 * 1024;  // Convert MB to bytes
+        return (long) volume.getMaxSizeMB() * 1024 * 1024;  // Convert MB to bytes, cast to long to prevent overflow
     }
 
     @Override
